Converted *Pixel to UIntPixel to resolve problems with reading in legacy raster files.
git-svn-id: http://Marc-PC/svn/Full@259 85f15e0e-5137-b144-a863-98bd542365a3

+++ trunk/biomass-succession/trunk/src/InitialBiomass.cs
@@ -84,7 +84,7 @@ namespace Landis.Extension.Succession.Biomass
         //  high 16-bits is the map code of the initial community and the
         //  low 16-bits is the ecoregion's map code
 
-        private static IDictionary<ushort, List<Landis.Library.AgeOnlyCohorts.ICohort>> sortedCohorts;
+        private static IDictionary<uint, List<Landis.Library.AgeOnlyCohorts.ICohort>> sortedCohorts;
         //  Age cohorts for an initial community sorted from oldest to
         //  youngest.  Key = initial community's map code
 
@@ -92,7 +92,7 @@ namespace Landis.Extension.Succession.Biomass
 
         //---------------------------------------------------------------------
 
-        private static uint ComputeKey(ushort initCommunityMapCode,
+        private static uint ComputeKey(uint initCommunityMapCode,
                                        ushort ecoregionMapCode)
         {
             return (uint)((initCommunityMapCode << 16) | ecoregionMapCode);
@@ -103,7 +103,7 @@ namespace Landis.Extension.Succession.Biomass
         static InitialBiomass()
         {
             initialSites = new Dictionary<uint, InitialBiomass>();
-            sortedCohorts = new Dictionary<ushort, List<Landis.Library.AgeOnlyCohorts.ICohort>>();
+            sortedCohorts = new Dictionary<uint, List<Landis.Library.AgeOnlyCohorts.ICohort>>();
         }
 
         //---------------------------------------------------------------------
